Move shared volume claim expansion features from Elasticsearch package to common
